refactor addressInput so clicking the login button will now save the entered address to local storage and retrieve from it on pageload
