+++ src/views/Login/Login.tsx
@@ -1,14 +1,14 @@
-import React, { useCallback, useMemo, useState } from "react";
+import React, { useCallback, useEffect, useMemo, useState } from "react";
 import Page from "components/Page";
 import ExistingUserDecideButtonGroup from "./components/ExistingUserDecideButtonGroup";
 import Logo from "components/Logo";
-import AddressInput from "./components/AddressInput";
 import OnboardingStepper from "./components/OnboardingStepper";
-import { Button, Checkbox, FormControlLabel, FormGroup } from "@mui/material";
-
-// import getAccount from "utils/api/getAccount";
+import { Autocomplete, Button, Checkbox, FormControlLabel, FormGroup, InputProps, TextField } from "@mui/material";
 import useLocalStorage from "hooks/useLocalStorage";
 import { NavLink } from "react-router-dom";
+
+// import getAccount from "utils/api/getAccount";
+
 /* 
   Component selection considerations (design)
 
@@ -35,10 +35,54 @@ import { NavLink } from "react-router-dom";
 
 */
 
+const autocompleteSx = {
+  padding: "16px",
+  minWidth: "200px",
+  maxWidth: "600px",
+  width: "100%",
+};
+
+export interface IInputOptions extends InputProps {
+  localStorageAccounts: Array<string>;
+  inputOnChangeFn: Function;
+}
+
+const AddressInput: React.FC<IInputOptions> = ({ localStorageAccounts, value, inputOnChangeFn }) => {
+  // const [value, setValue] = useState("");
+
+  return (
+    <Autocomplete
+      sx={autocompleteSx}
+      value={value}
+      freeSolo
+      disablePortal
+      onChange={(e, value) => inputOnChangeFn(value)}
+      options={localStorageAccounts}
+      renderInput={(params: any) => <TextField onChange={(e) => inputOnChangeFn(e.target.value)} {...params} label="Enter Account" />}
+    />
+  );
+};
+
 const Login: React.FC = () => {
-  const [isExistingUser, setIsExistingUser] = useState(false);
-  const [isRemembered, setIsRemembered] = useState(false);
+  const [isExistingUser, setIsExistingUser] = useState<boolean>(false);
+  const [isRemembered, setIsRemembered] = useState<boolean>(false);
   const [accounts, setAccounts] = useLocalStorage<Array<string>>("accounts", []); // stores user accounts in localStorage under "accounts" key
+  const [userInputAccount, setUserInputAccount] = useState<string>("");
+
+  const handleInputChange = useCallback((newValue: string | null) => {
+    if (newValue === null) {
+      return;
+    }
+
+    // TODO: Add formatting to address entry
+    // - Auto uppercase
+    // - 'JUP-' prefixing and hyphens at appropriate positions
+    //
+    // if (newValue.length > 2) {
+    //   newValue += "-";
+    // }
+    setUserInputAccount(newValue);
+  }, []);
 
   // keeps the dropdown menu options populated as localStorage changes
   const accountList = useMemo(() => {
@@ -67,15 +111,17 @@ const Login: React.FC = () => {
   // otherwise does nothing
   const handleLogin = useCallback(() => {
     if (isRemembered) {
-      const newAccounts = [...accounts, "test string"];
-      console.log("preparing to save newAccounts:", newAccounts);
-      setAccounts(newAccounts);
+      if (isValidAddress(userInputAccount)) {
+        const newAccounts = [...accounts, userInputAccount];
+        console.log("preparing to save newAccounts:", newAccounts);
+        setAccounts(newAccounts);
+        return;
+      }
+      console.error("invalid address entered");
+      // TODO: need to improve the behavior here, it shouldn't proceed to take them to the dashboard
     }
-  }, [isRemembered, setAccounts, accounts]);
+  }, [isRemembered, setAccounts, accounts, userInputAccount]);
 
-  // I have the Input keeping track of its own state (value)
-  // I have a button outside of that component
-  // When I click the button, I want to do something with the Input's value
   return (
     <Page>
       <Logo />
@@ -84,7 +130,7 @@ const Login: React.FC = () => {
         <OnboardingStepper />
       ) : (
         <>
-          <AddressInput localStorageAccounts={accountList !== undefined ? accountList : ["No Accounts"]} />
+          <AddressInput inputOnChangeFn={handleInputChange} localStorageAccounts={accountList !== undefined ? accountList : ["No Accounts"]} />
           <FormGroup>
             <FormControlLabel control={<Checkbox checked={isRemembered} onChange={handleRememberAccount} />} label="Remember Account?" />
             <NavLink to="/dashboard">
@@ -98,5 +144,20 @@ const Login: React.FC = () => {
     </Page>
   );
 };
+
+//
+// Helper Functions
+//
+
+// currently performs basic format checking, should be extended to support the JUP characters actually used in the NXT standards
+function isValidAddress(address: string) {
+  // TODO: confirm all letters get used, this currently validates for general structure but not any NXT/JUP standardization
+  const JUPREGEX = /^JUP-\w{4}-\w{4}-\w{4}-\w{5}$/;
+
+  if (address.match(JUPREGEX)) {
+    return true;
+  }
+  return false;
+}
 
 export default React.memo(Login);
