--- src/views/Login/components/AddressInput/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./AddressInput";
